requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
google-generativeai>=0.8.0
Pillow>=10.0.0
pdf2image>=1.16.0
PyMuPDF>=1.24.0
